JCLOUDS-3: Bootstrap nodes in a custom environment
Added the environment where the nodes in a group must be created in the
bootstrap data bag. This way the configuration is persisted and can be
used between runs, such as the node run list.

To accomodate the new bootstrap parameter, the method in the ChefService
used to create the data bag has been refactored and a the new
BootstrapConfig object has been created. This way, we'll be able to
extend the information in the bootstrap databag in the future without
changing the ChefService interface.