JCLOUDS-245/JCLOUDS-254: Fix live tests
Refactored the domain model to be immutable and addressed some
inconsistences with the Chef Server API model.

Removed all HEAD methods, as they have been removed from newer versions
of Chef. They were used to test the existance of a given resource, and
with newer versions the only way to do that is via a GET operation.

Now all live tests are passing for Community Chef 0.10.8, 11.0.6 and
Enterprise Chef.